fix(retrieve): gate unified retrieve behind sources feature
Move retrieve_unified + UnifiedHit + HitKind into a sub-module
gated by #[cfg(feature = "sources")]. Restores the
--no-default-features --features "cli server" build broken in
Task 5.
